cafe-fausse-backend/requirements.txt
Flask
Flask-CORS
psycopg2-binary
PyJWT
python-dotenv
gunicorn